requirements.txt
Flask
mysql-connector-python

#ativar a máquina virtual - venv\Scripts\activate
